Add screen to test LazyColumn with scope-keys and refactor screens to isolated files
